Fail-tests and clean-ups.
The concept of fail-tests was added within the standard CMake testing
routines.  The fail-test is a program which produces errors during
the compilation.  Mainly it is used to test that a certain code
prouces a given number of errors.
To control the errors, we introduce a `eq-error' directive which
can be used inside the source code in a comment on the line where
error happens.  The text followed by the directive is matched against
the output of the compiler.