Adding xfile interface; rename expand.h --> eq.h
In order to get a string representation of the tree one can
use pretty-print routines which were originally written to
output the text in the file.  For error-messaging one might
find it useful to get a representation of a certain tree
in memory.  As there is no portable way to do that, we
wrap a `FILE *' interface with `xfile *' which also allows
printf family functions within the memory.

Renaming of expand.h to eq.h was necessary because expand
is a historical name which does not have anything to do with
the current project.