Documentation for Eq project.
This is a first commit of the documentation system of the Eq project.  As a
main documentation source we use texinfo files.  The format itself is
suitable for producing pdf and html version.  In order to get man-page
we use a perl-script, which build a POD file, and makes a manual out of it.

HTML version uses a custom css -- eq.css and a jquery-based fix  to unify the
look of PDF and HTML versions.

Currently there is almost no useful information inside the file, it mainly
demonstrates that the system works.  So, if you go to docs/manual and
type make, you will get the following relevant files in the
docs/manual/build -- eq.pdf; eq.html; eq.1
they can be moved later to a more appropriate location.