Towards implementing optional SSA
The idea is that it is possible to set -DSSA=ON/OFF option for the
compiler.  In the first case static single assignment is enabled, and
variable redefinition will cause error.  Otherwise, if SSA is OFF, in
case redefinition, variable name will be substituted for a new one:
"<old_var_name> + {0, 1, ... }"

Currently the base mechanism should work, but there are still major
issues exist:

- Test sets differ for different compiler builds. We need to improve our
make system in order to build several versions and to have a separate
test set for each build.

- I improved memory tests in order to not only catch memory leaks, but
invalid reads/writes as well. This allowed to detect more memory-related
errors in the current test set. We need to fix these bugs, but it seems
tricky and I don't know how to do this at the moment.

- An introduction of SSA functionality caused new memory-related bugs
which seem difficult to fix at the moment. Perhaps we need to fix all
other memory-related bugs mentioned in the previous paragraph at first.
May be we can use another valgrind-like tool for resolving this?

- Because of the reason mentioned above I temporarily removed `md5'
tests from the test set, because it causes a heavy compiler break and it
doesn't allow test execution to finish.

- By the way, issue #3 is fixed. The cause was an unhandled error. I
added a test multiassign.tex which checks this case.