Fixing a parser Invalid read by valgrind.
As the memory tests show invalid reads now, one has to investigate where do
they come from.  Normally it is not that difficult, like in the example
above, when the token was accessed afeter being freed.  This is a general
pattern in the messages like that.

In case of token_location -- it is a well known feature, and it should be
fixed in all the parser.c.  I didn't look closely in the typechecker and
friends, but I would assumet that the same principle may be applied.

As for alternative tools to valgrind -- one can try out static verification
of clang.  It may catch some stuff as well, however, valgrind seems to be
the most advanced of all the existing tools.